+++ contracts/tests/resolvers/TestASValueResolver.sol
@@ -22,7 +22,7 @@ contract TestASValueResolver is ASResolver {
         address, /* recipient */
         bytes calldata, /* schema */
         bytes calldata, /* data */
-        uint256, /* expirationTime */
+        uint32, /* expirationTime */
         address /* msgSender */
     ) external payable virtual override returns (bool) {
         return msg.value == _targetValue;
